Update form styling and layout
Each layer of the form was wrapped in a div to achieve a proper grid
layout. Inside of each form layer, each label and input pair is also
wrapped in a div to achieve a proper flex layout and to maximize
accessibility (wrapping input elements in label elements seems to be bad
practice).