Add logic to toggle book read status
A function within the book object can be used to toggle the book's read
value.

The way in which form data was processed was changed to avoid type
coercion issues. Previously, it was passing booleans as strings, causing
all values to initially be toggled to false due to type coercion (all)
strings are truthy, so !("false") returns false.

Further, values that were not inputted by the user are now stored as
null.